requirements.txt
Django==5.2
djangorestframework==3.16.0
python-dotenv==1.0.0
requests==2.32.3 